+++ src/app/dataset-view/additional-components/metadata-component/components/add-polling-source/steps/preprocess-step/preprocess-step.component.spec.ts
@@ -12,11 +12,11 @@ import {
 } from "../../../set-transform/mock.data";
 import { QueriesSectionComponent } from "../../../set-transform/components/queries-section/queries-section.component";
 import { EngineSectionComponent } from "../../../set-transform/components/engine-section/engine-section.component";
-import { MonacoEditorModule } from "ngx-monaco-editor-v2";
 import { MatIconModule } from "@angular/material/icon";
 import { SharedTestModule } from "src/app/common/shared-test.module";
 import { ChangeDetectionStrategy } from "@angular/core";
 import { emitClickOnElementByDataTestId } from "src/app/common/base-test.helpers.spec";
+import { EditorModule } from "../../../../../../../shared/editor/editor.module";
 
 describe("PreprocessStepComponent", () => {
     let component: PreprocessStepComponent;
@@ -30,9 +30,9 @@ describe("PreprocessStepComponent", () => {
                 ApolloModule,
                 ApolloTestingModule,
                 FormsModule,
-                MonacoEditorModule.forRoot(),
                 MatIconModule,
                 SharedTestModule,
+                EditorModule,
             ],
             providers: [Apollo],
         })