+++ src/app/shared/editor/models/error.model.ts
@@ -1,0 +1,8 @@
+import { MarkerSeverity } from "monaco-editor";
+
+export interface EditorError {
+    line?: number;
+    col?: number;
+    severity: MarkerSeverity;
+    message: string;
+}