+++ CHANGELOG.md
@@ -9,10 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added initial loading screen for app
 - Added a progress bar when initializing the editor on the Data 
+- Showing editor markers on SQL query errors
 - Displayed lineage graph for a single dataset
 ### Changed
 - Separated main data query and lineage data 
 - Improve reporting of fatal errors from modal window to toaster window
+- Monaco-based editors isolated in `EditorModule`
 ### Fixed
 - GraphQL caching issues when switching between History and Lineage tabs
 - Correct author avatars on history tab